requirements.txt
blinker==1.8.2
Bootstrap-Flask==2.4.1
click==8.1.7
dnspython==2.7.0
dominate==2.9.1
email_validator==2.2.0
Faker==30.8.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==3.0.4
WTForms==3.2.1
